web/requirements.txt
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.12.4
django-filter==21.1
virtualenv==20.10.0
drf-yasg==1.20
psycopg2-binary==2.9.3
gunicorn==20.1.0
mozilla-django-oidc==2.0.0
django-tables2==2.4.1
tablib==3.2.0
pyyaml==6.0
pytz==2022.1
google-cloud-secret-manager==2.11.1
google-cloud-tasks==2.7.1
django-model-utils
ratelimit
backoff
requests-mock
